CustomFieldAdminPlugin: Adding L10N to the plugin.
Based on patch by g1itch including Russian translation. Thanks also to hasienda for helpful advice.

Closes #7738.



git-svn-id: https://trac-hacks.org/svn@11168 7322e99d-02ea-0310-aa39-e9a107903beb

+++ customfieldadminplugin/0.11/customfieldadmin/tests/web_ui.py
@@ -25,7 +25,8 @@ class CustomFieldAdminPageTestCase(unittest.TestCase):
         self.plugin = CustomFieldAdminPage(self.env)
 
     def tearDown(self):
-        self.env.destroy_db()
+        if hasattr(self.env, 'destroy_db'):
+            self.env.destroy_db()
         del self.env
 
     def test_systeminfo(self):
